cleaned and moved vanilla ElGamal code and example
